TODO list Int32/64 types Some execute fixes

+++ include/cpu.h
@@ -4,10 +4,12 @@
 #define	MEMORY_SIZE 20
 #define	REGISTER_COUNT 35
 
+#include <stdint.h>
+
 typedef struct 
 {
-    int	memory[MEMORY_SIZE];
-    long registers[REGISTER_COUNT];
+    int32_t memory[MEMORY_SIZE];
+    int32_t registers[REGISTER_COUNT];
 } CPU;
 
 #endif	/* CPU_H */
