+++ include/registers.h
@@ -1,6 +1,8 @@
 #ifndef	REGISTERS_H
 #define	REGISTERS_H
 
+#include <stdint.h>
+
 #include "cpu.h"
 
 #define	REG_HI 32
@@ -8,7 +10,7 @@
 #define	REG_PC 34
 
 void	registers_init(CPU *cpu);
-long	register_read(CPU *cpu, int id);
-void	register_write(CPU *cpu, int id, long value);
+int32_t	register_read(CPU *cpu, int id);
+void	register_write(CPU *cpu, int id, int32_t value);
 
 #endif	/* REGISTERS_H */
