+++ src/registers.c
@@ -1,3 +1,4 @@
+#include <stdint.h>
 #include <stdio.h>
 
 #include "registers.h"
@@ -8,9 +9,9 @@ void registers_init(CPU *cpu)
 		cpu->registers[i] = 0;
 }
 
-long register_read(CPU *cpu, int id)
+int32_t register_read(CPU *cpu, int id)
 {
-	long value = 0;
+	int32_t value = 0;
 
 	if(id >= 0 && id < REGISTER_COUNT)
 		value = cpu->registers[id];
@@ -20,7 +21,7 @@ long register_read(CPU *cpu, int id)
 	return(value);
 }
 
-void register_write(CPU *cpu, int id, long value)
+void register_write(CPU *cpu, int id, int32_t value)
 {
 	if(id > 0 && id < REGISTER_COUNT)
 		cpu->registers[id] = value;
